# qa-shop/api/openapi.yaml
openapi: 3.0.3
...
info:
  title: QA Shop Practice API
  version: "1.0.0"
  description: |
    Test otomasyonu pratiği için gerçek bir e-ticaret API'si. Mock değil:
    veri kalıcı, işlemler transaction içinde, stok gerçekten düşüyor.

    **Sandbox anahtarı göndermezsen** demo verisine salt okunur bağlanırsın —
    `GET` uçları çalışır, yazma denemeleri `401` döner. İlk denemeyi
    kolaylaştırmak için böyle.

    **Kendi alanını açmak için:** `POST /sandbox` → dönen `apiKey` değerini
    `X-Sandbox-Key` başlığında gönder. Alan 7 gün sonra otomatik silinir.

    **Her test koşumundan önce** `POST /sandbox/reset` çağır — temiz durum
    otomasyonun en temel disiplinidir ve burada bir uç olarak hazır.

    Demo hesabı: `demo@qashop.test` / `Password123!`
  contact:
    name: LearnQA.dev
    url: https://learnqa.dev
servers:
  - url: http://localhost:4000
    description: Lokal Docker (docker compose up)
...
security:
  - SandboxKey: []
paths:
...
  /api/v1/products:
    get:
      tags: [Katalog]
      summary: Ürün listesi
      description: |
        Pasif ürünler varsayılan olarak GİZLİDİR (soft delete). `includeInactive=true`
        ile görünür — "silinen ürün katalogda çıkmamalı ama eski siparişte
        durmalı" kuralının test edilebilir hâli.

        Markası olmayan ürünler vardır (`brand` alanı `null`). Bu bilinçlidir:
        INNER JOIN ile LEFT JOIN farkını gösteren satırlar olmadan JOIN dersi
        anlatılamaz.
      parameters:
        - { name: page, in: query, schema: { type: integer, default: 1, minimum: 1 } }
        - { name: size, in: query, schema: { type: integer, default: 20, maximum: 100 } }
        - name: sort
          in: query
          schema: { type: string, enum: [price, name, created, sku], default: sku }
        - { name: order, in: query, schema: { type: string, enum: [asc, desc], default: asc } }
        - { name: category, in: query, schema: { type: string, example: tshirts } }
        - { name: brand, in: query, schema: { type: string, example: Nike } }
        - { name: minPrice, in: query, schema: { type: number, example: 200 } }
        - { name: maxPrice, in: query, schema: { type: number, example: 800 } }
        - { name: q, in: query, schema: { type: string, example: Blue } }
        - { name: includeInactive, in: query, schema: { type: boolean, default: false } }
      responses:
        "200":
          description: Sayfalanmış ürün listesi
          content:
            application/json:
              schema:
                allOf:
                  - $ref: "#/components/schemas/Page"
                  - type: object
                    properties:
                      items:
                        type: array
                        items: { $ref: "#/components/schemas/Product" }
        "400":
          description: Tanımsız `sort` değeri
          content:
            application/json:
              schema: { $ref: "#/components/schemas/Error" }
  /api/v1/products/{id}:
    get:
      tags: [Katalog]
      summary: Ürün detayı
      description: |
        `rating_avg` YALNIZCA onaylı yorumlardan hesaplanır; bekleyen yorum
        puana etki etmemelidir. Pasif ürün `404` döner.
      parameters:
        - { name: id, in: path, required: true, schema: { type: integer }, example: 1 }
      responses:
        "200":
          description: Ürün
          content:
            application/json:
              schema:
                type: object
                properties:
                  product: { $ref: "#/components/schemas/ProductDetail" }
        "400": { description: id sayı değil }
        "404": { description: Ürün yok ya da artık satışta değil }
  /api/v1/products/{id}/variants:
    get:
      tags: [Katalog]
      summary: Ürünün varyantları ve stoğu
      description: |
        `available = stock_qty - reserved_qty`. Yalnızca `stock_qty`'ye bakmak
        sepette bekleyen adedi görmez ve oversell üretir.
      parameters:
        - { name: id, in: path, required: true, schema: { type: integer }, example: 1 }
      responses:
        "200":
          description: Varyant listesi
          content:
            application/json:
              schema:
                type: object
                properties:
                  productId: { type: integer }
                  total: { type: integer }
                  variants:
                    type: array
                    items: { $ref: "#/components/schemas/Variant" }
        "404": { description: Ürün bulunamadı }
...
  /api/v1/categories/{id}/products:
    get:
      tags: [Katalog]
      summary: Kategorideki ürünler (alt kategoriler dahil)
      description: |
        Üst kategori istendiğinde altındaki tüm kategorilerin ürünleri döner.
        Yalnızca doğrudan eşleşmeye bakan bir sorgu üst kategorilerde her zaman
        boş liste verirdi.
      parameters:
        - { name: id, in: path, required: true, schema: { type: integer } }
        - { name: page, in: query, schema: { type: integer, default: 1 } }
        - { name: size, in: query, schema: { type: integer, default: 20 } }
      responses:
        "200":
          description: Sayfalanmış ürün listesi
        "404": { description: Kategori bulunamadı }
...
  /api/v1/search:
    get:
      tags: [Katalog]
      summary: Ürün arama
      description: |
        2 karakterden kısa terim `400` döner. Boş aramayı sessizce "her şeyi
        döndür"e çevirmek, arayüzde fark edilmeyen bir performans sorununa
        dönüşür.
      parameters:
        - { name: q, in: query, required: true, schema: { type: string, minLength: 2 }, example: Dress }
        - { name: page, in: query, schema: { type: integer, default: 1 } }
        - { name: size, in: query, schema: { type: integer, default: 20 } }
      responses:
        "200": { description: Arama sonuçları }
        "400": { description: Terim çok kısa }

  # ─────────────────────────────────── Sepet ─────────────────────────────────
...
components:
  securitySchemes:
    SandboxKey:
      type: apiKey
      in: header
      name: X-Sandbox-Key
      description: |
        Hangi veri alanına bağlanacağını belirler. `POST /api/v1/sandbox` ile
        alınır. Gönderilmezse demo verisine SALT OKUNUR bağlanılır.
    BearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
      description: |
        `POST /api/v1/auth/login` ile alınır. Gerçek JWT formatındadır —
        jwt.io'ya yapıştırıp içindeki `sub`, `sandbox`, `jti`, `exp`
        alanlarını görebilirsin.
...
  schemas:
    Error:
      type: object
      description: |
        Hata gövdesi HER uçta aynı şekildedir. Her endpoint kendi hata şeklini
        döndürseydi test kodu her uç için ayrı yazılır ve kırılgan olurdu.
      properties:
        error:
          type: object
          properties:
            code: { type: string, example: OUT_OF_STOCK }
            message: { type: string, example: Stokta yalnızca 2 adet var }
            details:
              type: object
              nullable: true
              example: { requested: 5, available: 2 }
        correlationId:
          type: string
          example: req-8f3c1e2a-...
          description: Cevabın `X-Correlation-Id` başlığında da döner.
...
    Page:
      type: object
      properties:
        page: { type: integer, example: 1 }
        size: { type: integer, example: 20 }
        total: { type: integer, example: 111 }
        totalPages: { type: integer, example: 6 }
        hasNext: { type: boolean }
    Product:
      type: object
      properties:
        id: { type: integer }
        sku: { type: string, example: SKU-0001 }
        name: { type: string, example: Slim Fit White T-Shirt }
        description: { type: string, nullable: true }
        price: { type: number, example: 137.99 }
        currency: { type: string, example: TRY }
        is_active: { type: boolean }
        created_at: { type: string, format: date-time }
        category: { type: string, nullable: true, example: tshirts }
        category_name: { type: string, nullable: true }
        brand:
          type: string
          nullable: true
          description: Bilerek null olabilir — LEFT JOIN dersi için.
        variant_count: { type: integer, example: 3 }
        total_stock: { type: integer }
    ProductDetail:
      allOf:
        - $ref: "#/components/schemas/Product"
        - type: object
          properties:
            rating_avg:
              type: number
              nullable: true
              description: YALNIZCA onaylı yorumlardan hesaplanır.
            rating_count: { type: integer }
    Variant:
      type: object
      properties:
        id: { type: integer }
        sku: { type: string, example: SKU-0001-M }
        size: { type: string, enum: [S, M, L] }
        color: { type: string }
        price: { type: number }
        stock_qty: { type: integer }
        reserved_qty: { type: integer }
        available:
          type: integer
          description: stock_qty - reserved_qty. Sipariş edilebilir gerçek adet.
...